SM-61 moved authSlice to separate file
