+++ src/app/store/store.tsx
@@ -1,0 +1,9 @@
+import { configureStore } from '@reduxjs/toolkit';
+
+import { authSlice } from './authSlice';
+
+const store = configureStore({
+  reducer: { auth: authSlice.reducer },
+});
+
+export default store;
